uses observation-recorder
